+++ packages/web3/src/connect-button/demos/balance.tsx
@@ -15,6 +15,29 @@ const App: React.FC = () => {
           value: 1000000000000000000n,
         }}
       />
+      <ConnectButton
+        account={{
+          address: '3ea2cfd153b8d8505097b81c87c11f5d05097c18',
+          name: 'test.eth',
+        }}
+        balance={{
+          symbol: 'ETH',
+          decimals: 18,
+          value: 300000000000000000n,
+          coverAddress: false,
+        }}
+      />
+      <ConnectButton
+        account={{
+          address: '3ea2cfd153b8d8505097b81c87c11f5d05097c18',
+        }}
+        balance={{
+          symbol: 'ETH',
+          decimals: 18,
+          value: 1340000000000000000n,
+          coverAddress: false,
+        }}
+      />
     </Space>
   );
 };